Add Biome configuration and update package.json
This commit includes the addition of the Biome configuration file (biome.xml and biome.json) to set the Biome settings for the project. A new script "format" has been added to package.json, and Biome has been added to the project dependencies. The package lock file has been updated to include the new Biome dependency.