Implement Tailwind CSS configuration
The commit introduces Tailwind CSS into the project along with required dependencies such as autoprefixer and postcss. Necessary files for Tailwind configuration like 'tailwind.config.ts' and 'postcss.config.mjs' are added. Changes also include the import of the Tailwind CSS in 'root.tsx' and the addition of skipLibCheck and reordering in 'tsconfig.json'.
The local Tailwind CSS file has been removed and replaced with a CSS file in the '@suddenly-giovanni/ui' package. The 'tailwind.config.ts' file in the UI package no longer has a prefix set. Also, 'package.json' has been updated to output the produced CSS file in the correct directory, and a new CSS class has been added to the button component.