Refactor UI components and export adjustments
This commit includes refactoring of how children and className props are handled in the render functions of certain layout components. The data-testid attribute in the Button and SocialIcon components has also been corrected or added. Furthermore, the refactor expands on the data being exported from the index.tsx file to be more specific and organized.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>