Update module root manager settings
The module root manager settings have been updated to enhance the program execution. Key modifications consist of the addition of the "inherit-compiler-output" attribute and inclusion of several folders (mostly node_module ones) into the exclude list. This improvement streamlines the performance by reducing the number of files to be processed during runtime.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>