Add theme support to Storybook
This commit introduces the '@storybook/addon-themes' package to the project, enabling theme support for Storybook. The package has been added to .storybook/main.ts, and the theme configuration is added to .storybook/preview.ts. This update allows for theme switching, with light and dark themes available. Additionally, the versions of existing packages have been upgraded from 8.0.0-beta.2 to 8.0.0-beta.3.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>