Update CSS classes in layout components
Multiple updates have been made to the CSS class names in the components. The footer, header, root, and main layout components have undergone revisions. The changes are mainly related to how elements are displayed on the screen, with a focus on their positioning and the space they occupy. These are aimed at optimizing the UI layout in various screen sizes.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>