Refactor mode-toggle component and add SVG icons
Moved the MoonIcon and SunIcon from the mode-toggle component into a separate icons component. The code for these icons was significantly simplified by using SVG path data, which led to improved readability and a small reduction in code size. Also, conditionally rendered the moon and sun icons based on the theme state.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>