Add Shantell Sans font and apply to SuddenlyGiovanni component
This update includes the preconnection to Google Fonts and the addition of the Shantell Sans font to the root file. It also interchanges the h1 definition in the SuddenlyGiovanni component to utilize this new font. The specific font settings are detailed in the styles.css file.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>

Add preview-head.html for Storybook

This commit introduces a new file, preview-head.html, to serve static files in the Storybook. This includes a preconnect link to Google Fonts and a stylesheet link to load custom font weights and styles.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>