Refactor resume experience Route
The "experience.tsx" file has been removed and replaced with the "experiences.tsx" file in the resume route with added features. The change also updates the "route.tsx" file to utilize the new file. The main action involved is the restructuring of the experience section, aimed at improving the clarity and maintaining a more consistent style across the component.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>