Update styles and tags in resume header
The HTML tags for the summary and PDF download link in the resume header have been updated. The 'prose prose-slate' class has also been added to the hgroup for styling. Now, the summary is enclosed within a blockquote for consistency and the download link appears in a muted style for better clarity and design.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>