Refactor Languages component and adjust footer layout
Updated the Languages component of the resume routes to display languages and fluency levels using a table instead of a definition list. This enhances readability and data organization. Also, made a small adjustment to the footer layout to ensure full width utilization and modified the 'About Me' link to redirect to the homepage.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>