Add '@radix-ui/react-separator' to UI package
A new dependency, '@radix-ui/react-separator' version 1.0.3, has been added to the UI package. This has also been added to the pnpm lock file and used to create a new Separator component in the '/ui' directory, providing increased functionality and styling options in the User Interface. This change exports the Separator in 'src/index.tsx'.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>