Replace div with Separator component in resume sections
The div tag used for dividing elements in the education and experiences sections of the resume has been replaced with a Separator component for cleaner code. Import for the Separator component has been added in the affected files. Additionally, a minor change was made in the skills.tsx file to specify the return type as ReactElement.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>