Add accordion toggle feature to education route
Implemented a toggle feature to the education section of the resume. The toggle button allows for all education entries to be expanded or collapsed at once. This modification enhances user experience by providing a greater ease of navigation through the education history. It uses React hooks such as useState, useMemo, and useCallback for state management and optimization.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>