Update initial state in resume sections
This commit adds an update to the 'education', 'experience', and 'skills' sections on the resume page. The initial state was previously an empty array, but it is now updated to the first item in the list if it exists, otherwise, it defaults to an empty array. This change can ensure that there is a default value displayed when navigating to these sections.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>