Update React and React-DOM to newer version
The React and React-DOM versions in various dependencies across the project have been updated. This includes changes in the web app's package.json, UI package's package.json and pnpm-lock.yaml file. Although some of the version specifications seemed quite complex, such as multiple names for what seems to be the same dependency, the main action was changing the version number.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>