Adjust link styles in resume sections
Changes have been made to adjust the link styles in the "education" and "experiences" sections of the resume. Specifically, the margin and size of the icon associated with the link in the "education" section were updated. A slight change in margin was also applied to the link in the "experiences" section.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>