Add new Motivations route
A new route named Motivations has been added to the web app. This route includes a page titled 'My Motivations' that elaborates on the motivations, career goals, and personal preferences of a fictional character. The details on this page can be a guideline for prospective employers or team members about the working style and expectations of the character.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>