Update deployment process and Docker configuration
This commit modifies the Github action deploy script, moves Dockerfile, and relocates fly.toml to a new directory, ensuring an efficient Docker setup. The app's build-and-run instructions are added in the README. A new docker-compose.yml file is created to help manage services. A .dockerignore file is removed from the root and a new one is added to the application's directory, reducing unnecessary build context. Unnecessary scripts build.sh and start.sh are removed. The Dockerfile is also enhanced to improve the container setup process.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>