Disable eslint for mapper.ts in resume route
For the working module, mapper.ts under the resume route, the eslint rule for 'no-explicit-any' is temporarily disabled. This change is due to needing quick fixes in a 'Work In Progress' module where the rules might be causing interruptions. The plan is to eventually adhere to the standard and reactivate the rule.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>