Allow typecheck to continue on test failure
This commit edits the behavior of the typecheck action within the push-checks workflow. Previously, a failure in typecheck would halt the workflow, but now it will continue to run even if an error is encountered. This change is intended to ensure that all tests are run regardless of individual failures, improving overall test coverage feedback.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>