feat(resume): refactor schema interfaces and structure
The schema interfaces for parts such as Resume, Basics, Language,
Interest, and Publication have been refactored for better structure.
Annotations for each field have been added to make the APIs more
descriptive. This enhancement should provide more clarity and precision
regarding the nature of the fields used in the interface and thus
improve the developer experience.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>