feat: Remove @conform-to/react dependency and refactor components
The "@conform-to/react" package is removed from the dependencies in
"pnpm-lock.yaml" and "package.json". This involves minor changes to
component files, particularly "header.tsx" which no longer uses the
functions from the removed package. Additionally, the character value
for the theme state in the "mode-toggle" component has been modified and
 its type is now exported.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>