feat(web/package.json): update @types/react and other dependencies
This commit updates the version of @types/react from 18.2.70 to 18.2.72
in the web package.json. Additionally, the dependencies of several
packages in pnpm-lock.yaml that rely on @types/react have been updated.
This is done to ensure compatibility and to benefit from bug fixes and
performance improvements in the latest version.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>