feat(seo-utilities): Remove Changesets workflow file
The Changesets workflow file in the SEO Utilities package was removed.
This workflow used to automate versioning and release notes generation
and was run on an Ubuntu environment. The removal includes the deletion
of all configuration code in the file, meaning actions such as code
checkout, caching, pnpm and node setup, dependency installation, and
release PR creation will no longer be automated.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>