feat(open-graph-protocol): refactor imports of Types
This commit changes the imports of the `Types` definitions across
multiple files in the `open-graph-protocol` package. We've moved the
definitions to a separate file (`types.ts`), to improve the
maintainability and coherence of the codebase. Now, rather than
importing from individual files, types are imported from a single,
static location.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>