feat(resume): Update type import paths and move schema files
This commit updates type import paths in various resume related
components to match the new location of schema files. The server.schema
files previously present in routes/resume directory are moved to a
dedicated schemas directory at the root level. This restructuring
simplifies the project structure and makes it easier to manage and
locate schema files.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>