fix(web app): Update env type in root.tsx
In the 'root.tsx' module of the web app, the type of the 'env' prop has
been updated. Instead of arbitrarily accepting a record of string keys
and string values, it now expects to match the structure of the 'ENV'
type. This provides enhanced type-checking and consistency across uses
of environmental variables.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>