styles(web-app): update styles and lint ignore in resume basics and experiences
The styles of two components in the resume module (basics and
experiences) have been updated to give paragraphs a class name. This
allows better control over the styling of these elements. Also, an
explanation for ignoring lint's "no-array-index-key" warning has been
added. Now, using the index as a key for mapped elements is explicitly
noted as intended behaviour.

Signed-off-by: Giovanni Ravalico <15946771+suddenlyGiovanni@users.noreply.github.com>