Update GitHub action setup and environment variables for package installation
Refactor the `setup` GitHub action and various workflow files to use a renamed input -- `github_packages_token` in place of `github_token`. This new token is intended for accessing the GitHub Packages registry. The .npmrc file and environment settings for the installation of dependencies have also been adjusted to align with this change.

Included within the update is the creation of a `.envrc` file to automatically populate environment variables using 'direnv' for a streamlined developer experience. Additionally, debugging steps have been introduced to output the contents of the repository and the .npmrc file to aid in troubleshooting.